+++ src/org/balote/geocodingapp/models/LatLngModel.java
@@ -2,8 +2,8 @@ package org.balote.geocodingapp.models;
 
 public class LatLngModel {
 
-	private long latitude;
-	private long longitude;
+	private double latitude;
+	private double longitude;
 	
 	public LatLngModel() {
 		
@@ -17,28 +17,28 @@ public class LatLngModel {
 	/**
 	 * @return the latitude
 	 */
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
 	/**
 	 * @return the longitude
 	 */
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
 	/**
 	 * @param longitude the longitude to set
 	 */
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
